Added JWT to determine user login status.添加了JWT用以判断用户登录状态

+++ src/main/java/com/share/controller/ArticleController.java
@@ -1,5 +1,8 @@
 package com.share.controller;
 
+import com.share.annotation.UserLoginInfo;
+import com.share.annotation.UserLoginToken;
+import com.share.exceptions.ShareException;
 import com.share.result.RestObject;
 import com.share.result.RestResponse;
 import com.share.ro.articleRo.AllEArticleRo;
@@ -21,29 +24,54 @@ public class ArticleController {
     @Autowired
     ArticleService articleService;
 
-    @ApiOperation("发布信息")
-    @PostMapping("/addArticle")
-    public RestObject<String> addArticle(@RequestBody ArticleVo articleVo){
-        articleVo.setCreateTime(LocalDateTime.now());
-        articleService.addArticle(articleVo);
-        return RestResponse.makeOKRsp("新增成功");
+    @UserLoginInfo
+    @ApiOperation("发布信息。url参数：userId；body参数：title,text")
+    @PostMapping("/addArticle/{userId}")
+    public RestObject<String> addArticle(@RequestBody ArticleVo articleVo,@PathVariable int userId){
+        String text = articleVo.getText();
+        String title = articleVo.getTitle();
+        if( text == null || title==null){
+            throw new ShareException("有必填项未填写");
+        }else {
+            if (text.equals("") || title .equals("")){
+                throw new ShareException("有必填项未填写");
+            }else {
+                articleVo.setUserId(userId);
+                articleVo.setCreateTime(LocalDateTime.now());
+                articleService.addArticle(articleVo);
+                return RestResponse.makeOKRsp("发布成功!");
+            }
+        }
     }
 
-    @ApiOperation("修改信息")
-    @PostMapping("/updateArticle/{id}")
-    public RestObject<String> updateArticle(@PathVariable int id, @RequestBody ArticleVo articleVo){
-        articleVo.setUpdateTime(LocalDateTime.now());
-        articleService.updateArticle(id,articleVo);
-        return RestResponse.makeOKRsp("修改成功");
+    @UserLoginToken
+    @ApiOperation("修改信息。url参数：userId，用户id、id，信息id。")
+    @PostMapping("/updateArticle/{id}/{userId}")
+    public RestObject<String> updateArticle(@PathVariable int id, @RequestBody ArticleVo articleVo,@PathVariable int userId){
+        ArticleRo articleRo = articleService.queryArticleById(id);
+        if (userId!=articleRo.getUserId()){
+            return RestResponse.UserErrRsp("你无权修改!");
+        }else {
+            articleVo.setUpdateTime(LocalDateTime.now());
+            articleService.updateArticle(id,articleVo);
+            return RestResponse.makeOKRsp("修改成功!");
+        }
     }
 
+    @UserLoginToken
     @ApiOperation("删除信息")
-    @PostMapping("/deleteArticle/{id}")
-    public RestObject<String> deleteArticle(@PathVariable int id){
-        articleService.deleteArticle(id);
-        return RestResponse.makeOKRsp("删除成功");
+    @PostMapping("/deleteArticle/{id}/{userId}")
+    public RestObject<String> deleteArticle(@PathVariable int id,@PathVariable int userId){
+        ArticleRo articleRo = articleService.queryArticleById(id);
+        if (userId!=articleRo.getUserId()){
+            return RestResponse.UserErrRsp("你无权删除!");
+        }else {
+            articleService.deleteArticle(id);
+            return RestResponse.makeOKRsp("删除成功!");
+        }
     }
 
+    @UserLoginToken
     @ApiOperation("查询所有有效信息")
     @GetMapping("/queryAllEArticle")
     public RestObject<List<AllEArticleRo>> queryAllArticle(){
@@ -51,12 +79,14 @@ public class ArticleController {
         return RestResponse.makeOKRsp(articleService.queryAllArticle());
     }
 
+    @UserLoginToken
     @ApiOperation("按用户id查询信息")
     @GetMapping("/queryArticleByUserId/{userId}")
     public RestObject<List<ArticleRo>> queryArticleByUserId(@PathVariable int userId){
         return RestResponse.makeOKRsp(articleService.queryArticleByUserId(userId));
     }
 
+    @UserLoginToken
     @ApiOperation("按id查询信息")
     @GetMapping("/queryArticleById/{id}")
     public RestObject<ArticleRo> queryArticleById(@PathVariable int id){
